requirements.txt
asgiref==3.2.3
Babel==2.8.0
certifi==2019.11.28
Django==3.0.3
django-address==0.2.1
django-phonenumber-field==4.0.0
mysqlclient==1.4.6
phonenumbers==8.11.5
Pillow==7.1.2
pytz==2019.3
sqlparse==0.3.0
